Рефакторинг модуля утилит, рефакторинг метода render, чтобы он мог принимать компоненты
