+++ src/utils/common.js
@@ -1,8 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 export const getRandomInteger = (min, max) => min + Math.floor(Math.random() * (max - min + 1));
 
 export const getFormatNumber = (num) => num.toString().padStart(2, `0`);
@@ -23,25 +18,3 @@ export const getFormatEditTime = (time) => {
 };
 
 export const getFormatText = (text) => text.toLowerCase().split(` `).join(`-`);
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
-
-export const renderTemplate = (container, template, place) => {
-  container.insertAdjacentHTML(place, template);
-};
-
-export const renderElement = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
