+++ src/view/day.js
@@ -1,5 +1,5 @@
 import AbstractView from './abstract.js';
-import {getFormatDate} from '../utils.js';
+import {getFormatDate} from '../utils/common.js';
 
 const createTripDaysItemElement = (day, index) => {
   const date = getFormatDate(day);
